Update script.js
Added data-id to be able to differentiate each card. Added function to hide cards if they don't match the value in the search bar. Commented-out the function that saves and loads the edited content, so then it won't affect the current results. Aiden will fix the saving and loading later.